Gate periodic reviews before idle agent spawn
